@@ -5,18 +5,18 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "@react-three/drei": "8.20.1",
-    "@react-three/fiber": "7.0.19",
+    "@react-three/drei": "9.23.5",
+    "@react-three/fiber": "8.2.2",
     "meshline": "^2.0.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-scripts": "5.0.1",
-    "three": "0.138.0"
+    "three": "0.143.0"
   },
   "devDependencies": {
-    "@types/three": "0.137.0",
-    "@types/react": "^17.0.26",
-    "@types/react-dom": "^17.0.9",
+    "@types/three": "0.143.0",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
     "babel-plugin-import": "^1.13.3",
