@@ -1,7 +1,7 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 
+import { createRoot } from 'react-dom/client'
 import './styles.css'
 import App from './App'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+createRoot(document.getElementById('root') as HTMLElement).render(<App />)
